requirements.txt
python-dotenv==1.1.1
polars==1.32.3
pandas==2.3.1
qdrant-client==1.15.0
fastembed==0.7.1
openai==1.107.1
scikit-learn==1.7.0
openpyxl==3.1.5
dash==3.2.0
dash-bootstrap-components==2.0.4

